:truck: Rename: Gathering component not to be shared in Write folder
다른데 쓸 것 같지 않은 걸 퍼뜨려놓으면 혼란이 생길 것 같아 페이지 하위 폴더에 모음
