+++ src/pages/Write/index.tsx
@@ -5,18 +5,18 @@ import remarkGfm from 'remark-gfm';
 import remarkRehype from 'remark-rehype';
 import rehypeReact from 'rehype-react/lib';
 import classNames from 'classnames/bind';
-import useCodemirror from '../hooks/useCodemirror';
+import useCodemirror from '../../hooks/useCodemirror';
 import styles from './Write.module.scss';
-import { ReactComponent as BoldIcon } from '../assets/markdown_bold.svg';
-import { ReactComponent as ItalicIcon } from '../assets/markdown_italic.svg';
-import { ReactComponent as StrikeIcon } from '../assets/markdown_strike.svg';
-import { ReactComponent as BlockquoteIcon } from '../assets/markdown_blockquote.svg';
-import { ReactComponent as LinkIcon } from '../assets/markdown_link.svg';
-import { ReactComponent as ImageIcon } from '../assets/markdown_image.svg';
-import { ReactComponent as CodeblockIcon } from '../assets/markdown_codeblock.svg';
-import { ReactComponent as BackIcon } from '../assets/back.svg';
-import { showToast } from '../components/Toast';
-import PublishModal from '../components/PublishModal';
+import { ReactComponent as BoldIcon } from '../../assets/markdown_bold.svg';
+import { ReactComponent as ItalicIcon } from '../../assets/markdown_italic.svg';
+import { ReactComponent as StrikeIcon } from '../../assets/markdown_strike.svg';
+import { ReactComponent as BlockquoteIcon } from '../../assets/markdown_blockquote.svg';
+import { ReactComponent as LinkIcon } from '../../assets/markdown_link.svg';
+import { ReactComponent as ImageIcon } from '../../assets/markdown_image.svg';
+import { ReactComponent as CodeblockIcon } from '../../assets/markdown_codeblock.svg';
+import { ReactComponent as BackIcon } from '../../assets/back.svg';
+import { showToast } from '../../components/Toast';
+import PublishModal from './PublishModal';
 
 let treeData: any;
 const cx = classNames.bind(styles);
